refactor: rename class to 'ConfigParser'
